[backend] unknwonStixCoreObjects api
